chore(plugin): refactor
