+++ plugin/package.json
@@ -11,13 +11,10 @@
   "scripts": {
     "start:datacatalog": "vite -c vite.config.datacatalog-web.ts",
     "start:sidebar": "vite -c vite.config.sidebar-web.ts",
-    "build": "run-s typecheck clean build:web build:plugin",
+    "build": "run-s typecheck clean build:plugin",
     "build:plugin": "run-s build:sidebar build:datacatalog",
-    "build:datacatalog": "vite build -c vite.config.datacatalog.ts",
-    "build:sidebar": "vite build -c vite.config.sidebar.ts",
-    "build:web": "run-s clean:web build:datacatalog-web build:sidebar-web",
-    "build:datacatalog-web": "vite build -c vite.config.datacatalog-web.ts",
-    "build:sidebar-web": "vite build -c vite.config.sidebar-web.ts",
+    "build:sidebar": "vite build -c vite.config.sidebar-web.ts; vite build -c vite.config.sidebar.ts",
+    "build:datacatalog": "vite build -c vite.config.datacatalog-web.ts; vite build -c vite.config.datacatalog.ts",
     "clean": "del dist",
     "clean:web": "del dist/web",
     "typecheck": "tsc",
