+++ plugin/src/sidebar.ts
@@ -1,6 +1,7 @@
 import { PostMessageProps } from "@web/extensions/sidebar/types";
 
 import html from "../dist/web/sidebar/index.html?raw";
+import modalHtml from "../dist/web/sidebar/index.html?raw";
 
 const reearth = (globalThis as any).reearth;
 
@@ -14,6 +15,8 @@ reearth.on("message", ({ action, payload }: PostMessageProps) => {
       type: action,
       payload: reearth.scene.captureScreen(),
     });
+  } else if (action === "modal") {
+    reearth.modal.show(modalHtml);
   }
 });
