--- plugin/web/extensions/sharedComponents/index.ts
@@ -1,26 +1,0 @@
-import Button from "@web/extensions/sharedComponents/Button";
-import Checkbox from "@web/extensions/sharedComponents/Checkbox";
-import Col from "@web/extensions/sharedComponents/Col";
-import Divider from "@web/extensions/sharedComponents/Divider";
-import Form from "@web/extensions/sharedComponents/Form";
-import Icon from "@web/extensions/sharedComponents/Icon";
-import Input from "@web/extensions/sharedComponents/Input";
-import { Content, Footer } from "@web/extensions/sharedComponents/Layout";
-import Row from "@web/extensions/sharedComponents/Row";
-import Space from "@web/extensions/sharedComponents/Space";
-import Typography from "@web/extensions/sharedComponents/Typography";
-
-export {
-  Button,
-  Checkbox,
-  Divider,
-  Form,
-  Icon,
-  Input,
-  Content,
-  Footer,
-  Row,
-  Col,
-  Space,
-  Typography,
-};
