+++ plugin/web/extensions/sidebar/components/content/MapSettings/index.tsx
@@ -1,6 +1,6 @@
-import { Checkbox, Row } from "@web/extensions/sharedComponents";
-import CommonPage from "@web/extensions/sidebar/components/Content/CommonPage";
+import CommonPage from "@web/extensions/sidebar/components/content/CommonPage";
 import { ReearthApi } from "@web/extensions/sidebar/types";
+import { Checkbox, Row } from "@web/sharedComponents";
 import { styled } from "@web/theme";
 import { memo } from "react";
