+++ plugin/web/extensions/sidebar/components/content/Share/index.tsx
@@ -1,5 +1,5 @@
-import { Row, Icon } from "@web/extensions/sharedComponents";
-import CommonPage from "@web/extensions/sidebar/components/Content/CommonPage";
+import CommonPage from "@web/extensions/sidebar/components/content/CommonPage";
+import { Row, Icon } from "@web/sharedComponents";
 import { styled } from "@web/theme";
 import { memo, useState } from "react";
