+++ plugin/web/extensions/storytelling/index.html
@@ -1,0 +1,20 @@
+<title></title>
+<style>
+    html, body{
+        background: transparent;
+    }
+    .root{
+        background: #fff;
+        border-radius: 0 0 8px 0;
+    }
+    html, body, .root {
+        width: 347px;
+        height: 800px;
+        transition: height 0.5s, width 0.5s;
+        border-radius: 0;
+    }
+</style>
+<body>
+    <div class="root" id="root"></div>
+</body>
+<script type="module" src="./main.tsx"></script>
