+++ plugin/web/extensions/storytelling/index.tsx
@@ -1,0 +1,5 @@
+const Storytelling: React.FC = () => {
+  return <div>Storytelling</div>;
+};
+
+export default Storytelling;
