+++ plugin/web/extensions/storytelling/main.tsx
@@ -1,0 +1,12 @@
+import Storytelling from "@web/extensions/storytelling";
+import ReactDOM from "react-dom/client";
+
+(async () => {
+  const element = document.getElementById("root");
+  if (element) {
+    const root = ReactDOM.createRoot(element);
+    root.render(<Storytelling />);
+  }
+})();
+
+export {};
