+++ plugin/web/sharedComponents/Icon/icons.ts
