+++ plugin/web/sharedComponents/Icon/index.tsx
