+++ plugin/web/sharedComponents/index.ts
@@ -1,0 +1,26 @@
+import Button from "@web/sharedComponents/Button";
+import Checkbox from "@web/sharedComponents/Checkbox";
+import Col from "@web/sharedComponents/Col";
+import Divider from "@web/sharedComponents/Divider";
+import Form from "@web/sharedComponents/Form";
+import Icon from "@web/sharedComponents/Icon";
+import Input from "@web/sharedComponents/Input";
+import { Content, Footer } from "@web/sharedComponents/Layout";
+import Row from "@web/sharedComponents/Row";
+import Space from "@web/sharedComponents/Space";
+import Typography from "@web/sharedComponents/Typography";
+
+export {
+  Button,
+  Checkbox,
+  Divider,
+  Form,
+  Icon,
+  Input,
+  Content,
+  Footer,
+  Row,
+  Col,
+  Space,
+  Typography,
+};
